WL#10162 Drop events for util tables
When dropping the util tables for upgrade, it's necessary to also drop
the events related to the table. This will enable new events to be
created which includes all columns of the table.

Make sure to add code to drop the extra event which may potentially
exist on mysql.ndb_schema due to legacy bug, although it was long ago
this is much clearer than the existing "allow_hardcoded_name" in
event_name_for_table(). Remove old workaround which is actually already
unreachable since the ndb_schema table can't be dropped any other way
than before upgrade.

+++ sql/ndb_util_table.cc
@@ -211,8 +211,7 @@ bool Ndb_util_table::define_table_add_column(
 }
 
 bool Ndb_util_table::create_table_in_NDB(
-    NdbDictionary::Table &new_table) const {
-
+    const NdbDictionary::Table &new_table) const {
   // Set correct database name on the Ndb object
   Db_name_guard db_guard(m_thd_ndb->ndb, m_db_name.c_str());
 
@@ -229,16 +228,38 @@ bool Ndb_util_table::drop_table_in_NDB(
     const NdbDictionary::Table &old_table) const {
   // Set correct database name on the Ndb object
   Db_name_guard db_guard(m_thd_ndb->ndb, m_db_name.c_str());
-
   NdbDictionary::Dictionary *dict = m_thd_ndb->ndb->getDictionary();
+
+  if (!drop_events_in_NDB())
+  {
+    push_warning("Failed to drop events for table '%s'",
+                 m_table_name.c_str());
+    return false;
+  }
+
   if (dict->dropTableGlobal(old_table) != 0) {
     push_ndb_error_warning(dict->getNdbError());
     push_warning("Failed to drop table '%s'", old_table.getName());
     return false;
   }
+
   return true;
 }
 
+bool Ndb_util_table::drop_event_in_NDB(const char *event_name) const {
+  NdbDictionary::Dictionary *dict = m_thd_ndb->ndb->getDictionary();
+  if (dict->dropEvent(event_name) != 0) {
+    if (dict->getNdbError().code == 4710 || dict->getNdbError().code == 1419) {
+      // Failed to drop event but return code says it was
+      // because the event didn't exist -> all ok
+      return true;
+    }
+    push_ndb_error_warning(dict->getNdbError());
+    push_warning("Failed to drop event '%s'", event_name);
+    return false;
+  }
+  return true;
+}
 
 bool Ndb_util_table::create() const {
   NdbDictionary::Table new_table(m_table_name.c_str());
